MovedHero section test to separate page
